+++ Conexión_Flask/Model/agentClass.py
@@ -35,7 +35,7 @@ import time
 import datetime
 import random
 
-from .pathfinder import Pathfinder
+from pathfinder import Pathfinder
 
 class RobotAgent(Agent):
     def __init__(self,name, model):
@@ -142,6 +142,9 @@ class RobotAgent(Agent):
             self.actionPoints -= 1
             self.model.actionsLog.append(('agent', self.idRobot, 'extinguish', ny, nx))
             self.model.actionsLog.append(('agent', self.idRobot, 'smoke', ny, nx))
+            
+            
+            
             print(f"[Agente {self.idRobot}] EXTINGUISH_FIRE→SMOKE en {(nx, ny)}, AP={self.actionPoints}")
             return True
         return False  
@@ -258,8 +261,7 @@ class RobotAgent(Agent):
             print(f"[Agente {self.idRobot}] No lleva víctima, no va a la salida")
             return
         
-        pathfinder = Pathfinder(self)
-        path, exit = pathfinder.closestExit()
+        path, exit = self.pathfinder.closestExit()
         print("Path:_", path)
         print("EXIT: ", exit)
 
@@ -300,14 +302,22 @@ class RobotAgent(Agent):
                 print("No se pudo mover a", (next_y, next_x))
                 break
 
-    # def damaged(self):
+    # def damaged(self):}
+
+    def estrategyActions(self):
+        self.pathfinder.closestPOI()
+        
 
     def step(self):
         # Reinicia PA y ejecuta hasta agotarlos
         self.actionPoints = 4
 
-        if self.model.randomStatus:
+        if self.model.randomStatus == True:
+            print("Entro a random to model")
             if self.carriesPOI:
                 self.saveVictim()
             else:
                 self.actions()
+        else:
+            print("Not RANDOM: ", self.model.randomStatus)
+            self.estrategyActions()